Move gas to its own module.
